Changed the data type of CapturePeriod to UInt32
The type was previously Int8 which caused an error when trying to write to the CapturePeriod index.

+++ Development/Objects/GXDLMSProfileGeneric.cs
@@ -450,7 +450,7 @@ namespace Gurux.DLMS.Objects
             }
             if (index == 4)
             {
-                return DataType.Int8;
+                return DataType.UInt32;
             }
             if (index == 5)
             {
